feature: refactor logging to provided for usage outside of the provider pipeline
