+++ lib/interfaces/plugins.js
@@ -20,6 +20,7 @@ var path = require('path')
 var express = require('express')
 var _ = require('lodash')
 const modulesWithKeyword = require('../modules').modulesWithKeyword
+const getLogger = require('../logging')
 
 module.exports = function (app) {
   return {
@@ -131,19 +132,16 @@ module.exports = function (app) {
 
   function handleMessageWrapper (app) {
     return (providerId, data) => {
-      if (app.config.settings.enablePluginLogging) {
-        var plugin = app.pluginsMap[providerId]
-        if (!_.isUndefined(plugin)) {
-          if (!plugin.logger) {
-            plugin.logger = new (require('../../providers/log'))({
-              app: app,
-              discriminator: providerId
-            })
-          }
-          plugin.logger.write(data)
-        } else {
-          debug(`unknown provider id: ${providerId}`)
+      var plugin = app.pluginsMap[providerId]
+      if (
+        !_.isUndefined(plugin) &&
+        app.config.settings.enablePluginLogging &&
+        plugin.loggingEnabled
+      ) {
+        if (!plugin.logger) {
+          plugin.logger = getLogger(app, providerId)
         }
+        plugin.logger(data)
       }
       app.handleMessage(providerId, data)
     }
@@ -171,6 +169,7 @@ module.exports = function (app) {
       debug('Starting plugin ' + pluginName)
       plugin.start(getPluginOptions(plugin.id).configuration, restart)
     }
+    plugin.loggingEnabled = options.loggingEnabled
     app.plugins.push(plugin)
     app.pluginsMap[plugin.id] = plugin
 
@@ -199,6 +198,7 @@ module.exports = function (app) {
         res.send('Saved configuration for plugin ' + plugin.id)
         plugin.stop()
         const options = getPluginOptions(plugin.id)
+        plugin.loggingEnabled = options.loggingEnabled
         if (options.enabled) {
           plugin.start(options.configuration, restart)
         }
